+++ CHAD/Assets/Music.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79e355b47b8ff4b6c8355b0c7e1a0627
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
