--- CHAD/Assets/Scripts/GameManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8379668b4e1f34d47a32df00513028e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
