+++ CHAD/Assets/Scripts/Managers/MusicManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b527940d5bacd490e948395989e911a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
