--- CHAD/Assets/Scripts/NetworkManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60f09f6c9f108484fa1cc7acb06ca2cc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
